API cleanup, add rtkPaginationUtil
