+++ src/core/rtkQuery/splitV3Api/filterApi.ts
@@ -1,9 +1,12 @@
-import { defaultSerializeQueryArgs } from '@reduxjs/toolkit/query';
-import { omit } from 'lodash';
-
+import {
+  paginatedForceRefetch,
+  paginatedQueryMerge,
+  serializePaginatedQueryArgs,
+  transformPaginatedResponse,
+} from '@/core/rtkPaginationUtil';
 import { splitV3Api } from '@/core/rtkQuery/splitV3Api';
 
-import type { InfiniteResultType, ListResultType, PaginationType } from '@/core/types/api';
+import type { InfiniteResultType, PaginationType } from '@/core/types/api';
 import type { CollectionGroupType } from '@/core/types/api/collection';
 import type { FilterType } from '@/core/types/api/filter';
 import type { SeriesType } from '@/core/types/api/series';
@@ -18,6 +21,7 @@ const filterApi = splitV3Api.injectEndpoints({
         },
       }),
     }),
+
     getFilteredGroupsInfinite: build.query<
       InfiniteResultType<CollectionGroupType>,
       PaginationType & { randomImages?: boolean, filterCriteria: FilterType }
@@ -28,30 +32,12 @@ const filterApi = splitV3Api.injectEndpoints({
         params,
         body: filterCriteria,
       }),
-      transformResponse: (response: ListResultType<CollectionGroupType>, _, args) => ({
-        pages: {
-          [args.page ?? 1]: response.List,
-        },
-        total: response.Total,
-      }),
-      // Only have one cache entry because the arg always maps to one string
-      serializeQueryArgs: ({ endpointDefinition, endpointName, queryArgs }) =>
-        defaultSerializeQueryArgs({
-          endpointName,
-          queryArgs: omit(queryArgs, ['page']),
-          endpointDefinition,
-        }),
-      // Always merge incoming data to the cache entry
-      merge: (currentCache, newItems) => {
-        const tempCache = { ...currentCache };
-        tempCache.pages = { ...currentCache.pages, ...newItems.pages };
-        return tempCache;
-      },
-      // Refetch when the page arg changes
-      forceRefetch({ currentArg, previousArg }) {
-        return currentArg !== previousArg;
-      },
+      transformResponse: transformPaginatedResponse,
+      serializeQueryArgs: serializePaginatedQueryArgs,
+      merge: paginatedQueryMerge,
+      forceRefetch: paginatedForceRefetch,
     }),
+
     getFilteredGroupSeries: build.query<
       InfiniteResultType<SeriesType>,
       { groupId: string, randomImages?: boolean, filterCriteria: FilterType }
